Refactor move flake parsing into adapter file
